Add ability to search by `--ray-id`
This functionality used to be available via another legacy endpoint
however it was deprecated[1] and removed in favour of a combination of
the `/received` endpoint and `jq` magic due to some ongoing performance
issues on the Cloudflare end.

Thankfully, the ability to search by Ray ID has returned on a new
dedicated endpoint[2] that hangs off the back of the `/received` route.

In order to use this, I've updated the client and the CLI to accept the
`--ray-id` parameter and build out the correct request.

[1]: https://github.com/cloudflare/logshare/issues/14#issuecomment-351995474
[2]: https://api.cloudflare.com/#logs-recieved-logs-rayids